requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.6.1
django-countries==7.5.1
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
psycopg2==2.9.6
boto3==1.28.10
django-storages==1.13.2

